requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
starlette>=0.27.0
uvicorn>=0.30.0
sse-starlette>=1.6.0
